requirements.txt
flask
flask-socketio
gspread
oauth2client
python-dotenv
eventlet
gunicorn
requests